requirements.txt
py-cord
requests
cryptography
python_dotenv
pillow
icecream